Tests/LibWeb/Text/expected/Editing/execCommand-indent.txt

<blockquote>
    <span>foobar</span>
    </blockquote><ul>
        <li>foo</li><ul>
        <li>bar</li>
        </ul><li>baz</li>
    </ul>
